+++ site/src/_data/artworks/index.js
@@ -1,0 +1,9 @@
+import fetchSheet from './fetchSheet.js';
+import { normalizeSheet } from './normalizeRecord.js';
+
+export async function loadArtworks(options = {}) {
+  const sheetResponse = await fetchSheet();
+  return normalizeSheet(sheetResponse, options);
+}
+
+export { fetchSheet, normalizeSheet };